- rt11626 - silly read() not-checking-return-value bug.

+++ server/dhcpd.c
@@ -34,7 +34,7 @@
 
 #ifndef lint
 static char ocopyright[] =
-"$Id: dhcpd.c,v 1.115.2.12 2004/07/10 00:11:18 dhankins Exp $ Copyright 2004 Internet Systems Consortium.";
+"$Id: dhcpd.c,v 1.115.2.13 2004/09/02 21:40:05 dhankins Exp $ Copyright 2004 Internet Systems Consortium.";
 #endif
 
   static char copyright[] =
@@ -536,22 +536,24 @@ int main (argc, argv, envp)
 	if ((i = open (path_dhcpd_pid, O_RDONLY)) >= 0) {
 		status = read (i, pbuf, (sizeof pbuf) - 1);
 		close (i);
-		pbuf [status] = 0;
-		pid = atoi (pbuf);
+		if (status > 0) {
+			pbuf [status] = 0;
+			pid = atoi (pbuf);
 
-		/* If the previous server process is not still running,
-		   write a new pid file immediately. */
-		if (pid && (pid == getpid() || kill (pid, 0) < 0)) {
-			unlink (path_dhcpd_pid);
-			if ((i = open (path_dhcpd_pid,
-				       O_WRONLY | O_CREAT, 0644)) >= 0) {
-				sprintf (pbuf, "%d\n", (int)getpid ());
-				write (i, pbuf, strlen (pbuf));
-				close (i);
-				pidfilewritten = 1;
-			}
-		} else
-			log_fatal ("There's already a DHCP server running.");
+			/* If the previous server process is not still running,
+			   write a new pid file immediately. */
+			if (pid && (pid == getpid() || kill (pid, 0) < 0)) {
+				unlink (path_dhcpd_pid);
+				if ((i = open (path_dhcpd_pid,
+					O_WRONLY | O_CREAT, 0644)) >= 0) {
+				    sprintf (pbuf, "%d\n", (int)getpid ());
+				    write (i, pbuf, strlen (pbuf));
+				    close (i);
+				    pidfilewritten = 1;
+				}
+			} else
+				log_fatal ("There's already a DHCP server running.");
+		}
 	}
 
 	/* If we were requested to log to stdout on the command line,
